requirements.txt
Django>=5.0.7,<5.1
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.9,<3
drf-spectacular>=0.27.2,<0.28
drf-yasg>=1.21.8,<1.22